backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
python-jose[cryptography] 
python-multipart
pydantic[email-validator]
pydantic_settings